+++ game/configedit.py
@@ -1,6 +1,5 @@
 import pbge
 from pbge import util
-import pygame
 import configparser
 
 default_config = configparser.ConfigParser()
@@ -35,11 +34,85 @@ class OptionToggler(object):
         )
 
 
-class ConfigEditor(object):
-    def __init__(self, predraw, dy=-125):
-        self.dy = dy
-        self.predraw = predraw
-        self.finished = False
+class ConfigEditor(pbge.widgetmenu.MenuWidget):
+    TAGS_TO_HIDE = {pbge.widgets.WTAG_TITLEMENU,}
+    TAGS_TO_DEACTIVATE = {pbge.widgets.WTAG_WIDGET,}
+    ACTIVATE_IMMEDIATELY = True
+
+    def __init__(self, dy=-125):
+        super().__init__(
+            -CONFIG_EDITOR_WIDTH//2, dy, CONFIG_EDITOR_WIDTH, 250, draw_border=True,
+            on_escape=self.save_and_quit
+        )
+
+        self.add_interior(pbge.widgets.LabelWidget(0,0,CONFIG_EDITOR_WIDTH,0,"General Options", font=pbge.BIGFONT))
+
+        OptionToggler.add_menu_toggle(self.scroll_column, "Fullscreen", "fullscreen", extra_fun=self.toggle_fullscreen)
+        OptionToggler.add_menu_toggle(self.scroll_column, "Stretch Screen", "stretchy_screen", extra_fun=self.toggle_stretchyscreen, section="ACCESSIBILITY")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Music On", "music_on", extra_fun=self.toggle_music)
+        volume_menu = pbge.widgetmenu.ColDropdownWidget(
+            CONFIG_EDITOR_WIDTH, "Music Volume", on_select=self.set_music_volume
+        )
+        for msg,val in self.VOLUME_LEVELS:
+            volume_menu.add_item(msg, None, data=val)
+        volume_menu.set_item_by_data(util.config.get("GENERAL", "music_volume"))
+        self.add_interior(volume_menu)
+        music_mode_menu = pbge.widgetmenu.ColDropdownWidget(
+            CONFIG_EDITOR_WIDTH, "Music Mode", on_select=self.set_music_mode, add_desc=True
+        )
+        music_mode_menu.add_item(pbge.MUSIC_MODE_PRELOAD, None, data=pbge.MUSIC_MODE_PRELOAD, desc="Preload all music files. This option takes the most memory and increases load time, but provides flawless playback.")
+        music_mode_menu.add_item(pbge.MUSIC_MODE_CACHED, None, data=pbge.MUSIC_MODE_CACHED, desc="Cache the most recently used music files. This option uses a medium amount of memory, but reduces interruptions when new music is loaded.")
+        music_mode_menu.add_item(pbge.MUSIC_MODE_STREAM, None, data=pbge.MUSIC_MODE_STREAM, desc="Stream music directly from disk. This option uses the least amount of memory, but causes a slight delay when the music stream is changed.")
+        music_mode_menu.set_item_by_data(util.config.get("GENERAL", "music_mode"))
+        self.add_interior(music_mode_menu)
+
+        OptionToggler.add_menu_toggle(self.scroll_column, "Sound FX On", "sound_on")
+        volume_menu = pbge.widgetmenu.ColDropdownWidget(
+            CONFIG_EDITOR_WIDTH, "Sound FX Volume", on_select=self.set_sound_volume
+        )
+        for msg,val in self.VOLUME_LEVELS:
+            volume_menu.add_item(msg, data=val)
+        volume_menu.set_item_by_data(util.config.get("GENERAL", "sound_volume"))
+        self.add_interior(volume_menu)
+
+        OptionToggler.add_menu_toggle(self.scroll_column, "Names Above Heads", "names_above_heads")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Auto Save on Scene Change", "auto_save")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Can Replay Adventures", "can_replay_adventures")
+        OptionToggler.add_menu_toggle(self.scroll_column, "No Escape From Main Menu", "no_escape_from_title_screen")
+        window_menu = pbge.widgetmenu.ColDropdownWidget(
+            CONFIG_EDITOR_WIDTH, "Window Size", on_select=self.set_window_size
+        )
+        for res in self.WINDOW_SIZES:
+            window_menu.add_item(res, data=res)
+        window_menu.set_item_by_data(util.config.get("GENERAL", "window_size"))
+        self.add_interior(window_menu)
+
+        resolution_menu = pbge.widgetmenu.ColDropdownWidget(
+            CONFIG_EDITOR_WIDTH, "Fullscreen Resolution", on_select=self.set_resolution
+        )
+        for res in self.RESOLUTIONS:
+            resolution_menu.add_item(res, data=res)
+        myres = util.config.get("GENERAL", "fullscreen_resolution")
+        if myres not in self.RESOLUTIONS:
+            resolution_menu.add_item(myres, data=myres)
+        resolution_menu.set_item_by_data(myres)
+        self.add_interior(resolution_menu)
+
+        OptionToggler.add_menu_toggle(self.scroll_column, "Lancemates repaint their mecha", "lancemates_repaint_mecha")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Announce start of player turns", "announce_pc_turn_start")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Scroll to start of action library", "scroll_to_start_of_action_library")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Auto-center map cursor", "auto_center_map_cursor")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Mouse scroll at map edges", "mouse_scroll_at_map_edges")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Show numbers in pilot info", "show_numbers_in_pilot_info")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Dock tile info panel", "dock_tile_info_panel")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Show skills used in dialogue", "show_convo_skills")
+        OptionToggler.add_menu_toggle(self.scroll_column, "Indicate lance members on map", "indicate_lance_tiles")
+
+        self.add_interior(pbge.widgets.LabelWidget(0,0,CONFIG_EDITOR_WIDTH,0,"\nDifficulty Options", font=pbge.BIGFONT))
+        for op in util.config.options("DIFFICULTY"):
+            OptionToggler.add_menu_toggle(self.scroll_column, op, op, section="DIFFICULTY")
+
+        self.add_interior(pbge.widgets.LabelWidget(0,0,CONFIG_EDITOR_WIDTH,0,"Save and Exit", on_click=self.save_and_quit))
 
     def toggle_fullscreen(self):
         # Actually toggle the fullscreen.
@@ -97,7 +170,10 @@ class ConfigEditor(object):
                 pbge.my_state.resume_music()
 
     def save_and_quit(self, *_args):
-        self.finished = True
+        # Export the new config options.
+        with open(util.user_dir("config.cfg"), "wt") as f:
+            util.config.write(f)
+        self.pop()
 
     WINDOW_SIZES = (
         "800x600", "1280x720", "1600x900", "1600x1200", "1920x1080", "2560x1440"
@@ -110,132 +186,30 @@ class ConfigEditor(object):
         ("50%", "0.5"), ("40%", "0.4"), ("30%", "0.3"), ("20%", "0.2"), ("10%", "0.1")
     )
 
-    def __call__(self):
-        self.column = pbge.widgets.ColumnWidget(
-            -CONFIG_EDITOR_WIDTH//2, self.dy, CONFIG_EDITOR_WIDTH, 250, draw_border=True
+
+class PopupGameMenu(pbge.widgetmenu.MenuWidget):
+    TAGS_TO_DEACTIVATE = {pbge.widgets.WTAG_WIDGET,}
+    ACTIVATE_IMMEDIATELY = True
+
+    def __init__(self):
+        super().__init__(
+            -150, -100, 300, 200, font=pbge.my_state.huge_font,
+            on_escape=self.exit_menu
         )
-        up_arrow = pbge.widgets.ButtonWidget(0, 0, 128, 16,
-                                             sprite=pbge.image.Image("sys_updownbuttons.png", 128, 16),
-                                             on_frame=0, off_frame=1)
-        down_arrow = pbge.widgets.ButtonWidget(0, 0, 128, 16,
-                                               sprite=pbge.image.Image("sys_updownbuttons.png", 128, 16),
-                                               on_frame=2, off_frame=3)
-        self.scroll_column = pbge.widgets.ScrollColumnWidget(
-            0, 0, CONFIG_EDITOR_WIDTH, 200 - 40, up_arrow, down_arrow, padding=5, focus_locked=True
-        )
-        self.column.add_interior(up_arrow)
-        self.column.add_interior(self.scroll_column)
-        self.column.add_interior(down_arrow)
+        _=self.add_item("Quit Game", self.do_quit)
+        _=self.add_item("Config Options", self.do_config)
+        _=self.add_item("Continue", self.exit_menu)
 
-        self.scroll_column.add_interior(pbge.widgets.LabelWidget(0,0,CONFIG_EDITOR_WIDTH,0,"General Options", font=pbge.BIGFONT))
+    def do_quit(self, _wid, _ev):
+        pbge.my_state.session_data[pbge.campaign.SDAT_GOT_QUIT] = True
+        self.pop()
 
-        OptionToggler.add_menu_toggle(self.scroll_column, "Fullscreen", "fullscreen", extra_fun=self.toggle_fullscreen)
-        OptionToggler.add_menu_toggle(self.scroll_column, "Stretch Screen", "stretchy_screen", extra_fun=self.toggle_stretchyscreen, section="ACCESSIBILITY")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Music On", "music_on", extra_fun=self.toggle_music)
-        volume_menu = pbge.widgets.ColDropdownWidget(
-            CONFIG_EDITOR_WIDTH, "Music Volume", on_select=self.set_music_volume
-        )
-        for msg,val in self.VOLUME_LEVELS:
-            volume_menu.add_item(msg, None, val)
-        volume_menu.my_menu_widget.menu.set_item_by_value(util.config.get("GENERAL", "music_volume"))
-        self.scroll_column.add_interior(volume_menu)
-        music_mode_menu = pbge.widgets.ColDropdownWidget(
-            CONFIG_EDITOR_WIDTH, "Music Mode", on_select=self.set_music_mode, add_desc=True
-        )
-        music_mode_menu.add_item(pbge.MUSIC_MODE_PRELOAD, None, pbge.MUSIC_MODE_PRELOAD, "Preload all music files. This option takes the most memory and increases load time, but provides flawless playback.")
-        music_mode_menu.add_item(pbge.MUSIC_MODE_CACHED, None, pbge.MUSIC_MODE_CACHED, "Cache the most recently used music files. This option uses a medium amount of memory, but reduces interruptions when new music is loaded.")
-        music_mode_menu.add_item(pbge.MUSIC_MODE_STREAM, None, pbge.MUSIC_MODE_STREAM, "Stream music directly from disk. This option uses the least amount of memory, but causes a slight delay when the music stream is changed.")
-        music_mode_menu.my_menu_widget.menu.set_item_by_value(util.config.get("GENERAL", "music_mode"))
-        self.scroll_column.add_interior(music_mode_menu)
+    def do_config(self, wid, _ev):
+        self.pop()
+        ConfigEditor.push_state_and_instantiate()
 
-        OptionToggler.add_menu_toggle(self.scroll_column, "Sound FX On", "sound_on")
-        volume_menu = pbge.widgets.ColDropdownWidget(
-            CONFIG_EDITOR_WIDTH, "Sound FX Volume", on_select=self.set_sound_volume
-        )
-        for msg,val in self.VOLUME_LEVELS:
-            volume_menu.add_item(msg, val)
-        volume_menu.my_menu_widget.menu.set_item_by_value(util.config.get("GENERAL", "sound_volume"))
-        self.scroll_column.add_interior(volume_menu)
-
-        OptionToggler.add_menu_toggle(self.scroll_column, "Names Above Heads", "names_above_heads")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Auto Save on Scene Change", "auto_save")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Can Replay Adventures", "can_replay_adventures")
-        OptionToggler.add_menu_toggle(self.scroll_column, "No Escape From Main Menu", "no_escape_from_title_screen")
-        window_menu = pbge.widgets.ColDropdownWidget(
-            CONFIG_EDITOR_WIDTH, "Window Size", on_select=self.set_window_size
-        )
-        for res in self.WINDOW_SIZES:
-            window_menu.add_item(res, res)
-        window_menu.my_menu_widget.menu.set_item_by_value(util.config.get("GENERAL", "window_size"))
-        self.scroll_column.add_interior(window_menu)
-
-        resolution_menu = pbge.widgets.ColDropdownWidget(
-            CONFIG_EDITOR_WIDTH, "Fullscreen Resolution", on_select=self.set_resolution
-        )
-        for res in self.RESOLUTIONS:
-            resolution_menu.add_item(res, res)
-        myres = util.config.get("GENERAL", "fullscreen_resolution")
-        if myres not in self.RESOLUTIONS:
-            resolution_menu.add_item(myres, myres)
-        resolution_menu.my_menu_widget.menu.set_item_by_value(myres)
-        self.scroll_column.add_interior(resolution_menu)
-
-        OptionToggler.add_menu_toggle(self.scroll_column, "Lancemates repaint their mecha", "lancemates_repaint_mecha")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Announce start of player turns", "announce_pc_turn_start")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Scroll to start of action library", "scroll_to_start_of_action_library")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Auto-center map cursor", "auto_center_map_cursor")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Mouse scroll at map edges", "mouse_scroll_at_map_edges")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Show numbers in pilot info", "show_numbers_in_pilot_info")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Dock tile info panel", "dock_tile_info_panel")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Show skills used in dialogue", "show_convo_skills")
-        OptionToggler.add_menu_toggle(self.scroll_column, "Indicate lance members on map", "indicate_lance_tiles")
-
-        self.scroll_column.add_interior(pbge.widgets.LabelWidget(0,0,CONFIG_EDITOR_WIDTH,0,"\nDifficulty Options", font=pbge.BIGFONT))
-        for op in util.config.options("DIFFICULTY"):
-            OptionToggler.add_menu_toggle(self.scroll_column, op, op, section="DIFFICULTY")
-
-        self.scroll_column.add_interior(pbge.widgets.LabelWidget(0,0,CONFIG_EDITOR_WIDTH,0,"Save and Exit", on_click=self.save_and_quit))
-
-        pbge.my_state.widgets.append(self.column)
-        keepgoing = True
-        while keepgoing and not self.finished and not pbge.my_state.got_quit:
-            ev = pbge.wait_event()
-            if ev.type == pbge.TIMEREVENT:
-                if self.predraw:
-                    self.predraw()
-                else:
-                    pbge.my_state.view()
-                pbge.my_state.do_flip()
-            elif ev.type == pygame.KEYDOWN:
-                if pbge.my_state.is_key_for_action(ev, "exit"):
-                    keepgoing = False
-
-        pbge.my_state.widgets.remove(self.column)
-
-        # Export the new config options.
-        with open(util.user_dir("config.cfg"), "wt") as f:
-            util.config.write(f)
-
-
-class PopupGameMenu(object):
-    def do_quit(self, enc_or_com):
-        enc_or_com.no_quit = False
-
-    def do_config(self, enc_or_com):
-        myconfigmenu = ConfigEditor(None)
-        myconfigmenu()
-
-    def __call__(self, enc_or_com):
-        my_menu = pbge.rpgmenu.Menu(-150, -100, 300, 200,
-                                   font=pbge.my_state.huge_font)
-        my_menu.add_item("Quit Game", self.do_quit)
-        my_menu.add_item("Config Options", self.do_config)
-        my_menu.add_item("Continue", False)
-        action = True
-        while action and enc_or_com.no_quit:
-            action = my_menu.query()
-            if action:
-                action(enc_or_com)
+    def exit_menu(self, _wid, _ev):
+        self.pop()
 
 
 def init_configedit():