+++ game/mechabrowser.py
@@ -13,7 +13,7 @@ class MSRPBlock(object):
         self.height = self.image.get_height()
 
     def render(self, x, y):
-        pbge.my_state.screen.blit(self.image, pygame.Rect(x, y, self.width, self.height))
+        _=pbge.my_state.screen.blit(self.image, pygame.Rect(x, y, self.width, self.height))
 
 
 class MechaBrowseIP(gears.info.InfoPanel):
@@ -36,19 +36,22 @@ class BrowserDesc(object):
         self.info_cache[item] = it
         return it
 
-    def __call__(self, menuitem):
+    def __call__(self, item):
         mydest = self.ITEM_INFO_AREA.get_rect()
-        item = menuitem.value
         if item:
             myinfo = self.get_browse_info(item)
             if myinfo:
-                myinfo.render(mydest.x, mydest.y)
+                _=myinfo.render(mydest.x, mydest.y)
 
 
-class MechaBrowser(object):
-    MENU_AREA = pbge.frects.Frect(0, -200, 350, 400)
+class MechaBrowser(pbge.widgetmenu.MenuWidget):
+    TAGS_TO_HIDE = {pbge.widgets.WTAG_TITLEMENU,}
+    TAGS_TO_DEACTIVATE = {pbge.widgets.WTAG_WIDGET,}
+    ACTIVATE_IMMEDIATELY = True
 
     def __init__(self):
+        super().__init__(0, -200, 350, 400, font=pbge.BIGFONT, on_escape=self.exit_browser)
+
         self.mek_list = list()
         for mek in gears.selector.DESIGN_LIST:
             if isinstance(mek,gears.base.Mecha):
@@ -61,16 +64,16 @@ class MechaBrowser(object):
                         numek.colors = gears.color.random_mecha_colors()
                 self.mek_list.append(numek)
 
-    def __call__(self, predraw):
-        mymenu = pbge.rpgmenu.Menu(self.MENU_AREA.dx, self.MENU_AREA.dy, self.MENU_AREA.w, self.MENU_AREA.h,
-                                   font=pbge.BIGFONT, predraw=predraw)
-        mymenu.descobj = BrowserDesc()
-
         for mek in self.mek_list:
-            mymenu.add_item(mek.get_full_name(),mek)
-        mymenu.sort()
-        mymenu.add_item("[exit]", None)
+            _=self.add_item(mek.get_full_name(),None, data=mek)
+        self.sort()
+        _=self.add_item("[exit]", self.exit_browser)
+        self.desc_obj = BrowserDesc()
 
-        while mymenu.query():
-            pass
+    def _render(self, delta):
+        self.desc_obj(self.current_data)
+        super()._render(delta)
+
+    def exit_browser(self, _wid, _ev):
+        self.pop()